Move shouldReturnSeriesWhenChannelRegistryIsUnavailableAndSeriesIsCached to FetchSerialPublicationByIdentifierAndYearHandlerTest
