+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/series/FetchSeriesByIdentifierAndYearHandlerTest.java
@@ -46,7 +46,7 @@ class FetchSeriesByIdentifierAndYearHandlerTest extends FetchSerialPublicationBy
     private static final String CHANNEL_REGISTRY_PATH_ELEMENT = "/findseries/";
 
     @Override
-    protected String getChannelRegistryPathParameter() {
+    protected String getChannelRegistryPathElement() {
         return CHANNEL_REGISTRY_PATH_ELEMENT;
     }
 
@@ -103,31 +103,17 @@ class FetchSeriesByIdentifierAndYearHandlerTest extends FetchSerialPublicationBy
         return new TestChannel(year, identifier, "Series");
     }
 
+    @Override
+    protected String getPath() {
+        return "series";
+    }
+
     @BeforeEach
     void setup() {
         this.handlerUnderTest = new FetchSeriesByIdentifierAndYearHandler(environment,
                                                                           this.channelRegistryClient,
                                                                           this.cacheService,
                                                                           super.getAppConfigWithCacheEnabled(false));
-    }
-
-    @Test
-    void shouldReturnRedirectWhenChannelRegistryReturnsRedirect() throws IOException {
-        var year = String.valueOf(randomYear());
-        var requestedIdentifier = UUID.randomUUID().toString();
-        var newIdentifier = UUID.randomUUID().toString();
-        var newChannelRegistryLocation = UriWrapper.fromHost(channelRegistryBaseUri)
-                                             .addChild(CHANNEL_REGISTRY_PATH_ELEMENT, newIdentifier, year)
-                                             .toString();
-        mockRedirectedClient(requestedIdentifier, newChannelRegistryLocation, year, CHANNEL_REGISTRY_PATH_ELEMENT);
-        handlerUnderTest.handleRequest(constructRequest(year, requestedIdentifier, MediaType.ANY_TYPE),
-                                       output,
-                                       context);
-        var response = GatewayResponse.fromOutputStream(output, HttpResponse.class);
-        assertEquals(HTTP_MOVED_PERM, response.getStatusCode());
-        var expectedLocation = createPublicationChannelUri(newIdentifier, SERIES_PATH, year).toString();
-        assertEquals(expectedLocation, response.getHeaders().get(LOCATION));
-        assertEquals(WILD_CARD, response.getHeaders().get(ACCESS_CONTROL_ALLOW_ORIGIN));
     }
 
     @Test
